+++ app/src/androidTest/java/com/geeksville/mesh/compose/MessageItemTest.kt
@@ -21,13 +21,13 @@ import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.geeksville.mesh.ui.common.preview.NodePreviewParameterProvider
 import com.geeksville.mesh.ui.message.components.MessageItem
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.meshtastic.core.database.model.Message
 import org.meshtastic.core.model.MessageStatus
+import org.meshtastic.core.ui.component.preview.NodePreviewParameterProvider
 
 @RunWith(AndroidJUnit4::class)
 class MessageItemTest {
