+++ app/src/main/java/com/geeksville/mesh/ui/node/components/NodeItem.kt
@@ -47,13 +47,13 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import com.geeksville.mesh.ConfigProtos.Config.DisplayConfig
-import com.geeksville.mesh.ui.common.components.SignalInfo
-import com.geeksville.mesh.ui.common.preview.NodePreviewParameterProvider
 import org.meshtastic.core.database.model.Node
 import org.meshtastic.core.database.model.isUnmessageableRole
 import org.meshtastic.core.model.util.toDistanceString
 import org.meshtastic.core.strings.R
 import org.meshtastic.core.ui.component.BatteryInfo
+import org.meshtastic.core.ui.component.SignalInfo
+import org.meshtastic.core.ui.component.preview.NodePreviewParameterProvider
 import org.meshtastic.core.ui.theme.AppTheme
 import org.meshtastic.feature.node.component.LastHeardInfo
 import org.meshtastic.feature.node.component.NodeChip
